refactor: move DataConverter to dedicated module
